server/requirements.txt
click==7.1.1
Flask==1.1.2
Flask-JWT-Extended==3.24.1
Flask-SQLAlchemy==2.4.1
Flask-Cors==3.0.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
PyJWT==1.7.1
six==1.14.0
SQLAlchemy==1.3.15
Werkzeug==1.0.1
Flask-Bcrypt==0.7.1
bcrypt==3.1.7
Cerberus==1.3.2
Flask-Testing==0.8.0
cffi==1.14.0
python-dateutil==2.8.1




